+++ src/index.ts
@@ -1,6 +1,6 @@
 import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 
-import instanceConfigFields from './instanceConfigFields';
+import { instanceConfigFields } from './instanceConfigFields';
 import { integrationSteps } from './steps';
 import { IntegrationConfig } from './types';
 import validateInvocation from './validateInvocation';